refactor: reorganize utility files
