+++ src/components/settings/sections/RAGSection.tsx
@@ -6,7 +6,7 @@ import SmartComposerPlugin from '../../../main'
 import { EmbeddingDbManageModal } from '../../../settings/EmbeddingDbManageModal'
 import { ExcludedFilesModal } from '../../../settings/ExcludedFilesModal'
 import { IncludedFilesModal } from '../../../settings/IncludedFilesModal'
-import { findFilesMatchingPatterns } from '../../../utils/globUtils'
+import { findFilesMatchingPatterns } from '../../../utils/glob-utils'
 import { ObsidianButton } from '../../common/ObsidianButton'
 import { ObsidianDropdown } from '../../common/ObsidianDropdown'
 import { ObsidianSetting } from '../../common/ObsidianSetting'
