+++ src/core/llm/gemini.ts
@@ -18,7 +18,7 @@ import {
   LLMResponseStreaming,
 } from '../../types/llm/response'
 import { LLMProvider } from '../../types/provider.types'
-import { parseImageDataUrl } from '../../utils/image'
+import { parseImageDataUrl } from '../../utils/llm/image'
 
 import { BaseLLMProvider } from './base'
 import {
