+++ src/utils/chat/parse-smtcmp-block.test.ts
