+++ src/utils/chat/parse-smtcmp-block.ts
